rename to local_proxy, move to config to module
