+++ compute_tools/src/compute.rs
@@ -34,7 +34,7 @@ use nix::sys::signal::{kill, Signal};
 use remote_storage::{DownloadError, RemotePath};
 
 use crate::checker::create_availability_check_data;
-use crate::config::configure_local_proxy;
+use crate::local_proxy;
 use crate::logger::inlinify;
 use crate::pg_helpers::*;
 use crate::spec::*;
@@ -888,8 +888,8 @@ impl ComputeNode {
         drop(client);
 
         if let Some(ref local_proxy) = spec.local_proxy_config {
-            info!("configuring local-proxy");
-            configure_local_proxy(local_proxy).context("apply_config configure_local_proxy")?;
+            info!("configuring local_proxy");
+            local_proxy::configure(local_proxy).context("apply_config local_proxy")?;
         }
 
         // Run migrations separately to not hold up cold starts
@@ -943,14 +943,14 @@ impl ComputeNode {
         }
 
         if let Some(ref local_proxy) = spec.local_proxy_config {
-            info!("configuring local-proxy");
+            info!("configuring local_proxy");
 
             // Spawn a thread to do the configuration,
             // so that we don't block the main thread that starts Postgres.
             let local_proxy = local_proxy.clone();
             let _handle = Some(thread::spawn(move || {
-                if let Err(err) = configure_local_proxy(&local_proxy) {
-                    error!("error while configuring localproxy: {err:?}");
+                if let Err(err) = local_proxy::configure(&local_proxy) {
+                    error!("error while configuring local_proxy: {err:?}");
                 }
             }));
         }
@@ -1043,14 +1043,14 @@ impl ComputeNode {
         }
 
         if let Some(local_proxy) = &pspec.spec.local_proxy_config {
-            info!("configuring local-proxy");
+            info!("configuring local_proxy");
 
             // Spawn a thread to do the configuration,
             // so that we don't block the main thread that starts Postgres.
             let local_proxy = local_proxy.clone();
             let _handle = thread::spawn(move || {
-                if let Err(err) = configure_local_proxy(&local_proxy) {
-                    error!("error while configuring localproxy: {err:?}");
+                if let Err(err) = local_proxy::configure(&local_proxy) {
+                    error!("error while configuring local_proxy: {err:?}");
                 }
             });
         }
